Document accessing SOAP client metrics and Vert.x HttpClient metrics
