+++ integration-tests/metrics/src/test/java/io/quarkiverse/cxf/metrics/it/MetricsTest.java
@@ -98,8 +98,6 @@ public class MetricsTest {
 
     }
 
-    // TODO test like in     io.quarkus.micrometer.deployment.binder.VertxHttpClientMetricsTest
-
     public static Map<String, Object> getMetrics() {
         final String body = RestAssured.given()
                 .header("Content-Type", "application/json")
