+++ integration-tests/metrics/src/test/java/io/quarkiverse/cxf/metrics/it/VertxHttpClientMetricsIT.java
@@ -1,0 +1,8 @@
+package io.quarkiverse.cxf.metrics.it;
+
+import io.quarkus.test.junit.QuarkusIntegrationTest;
+
+@QuarkusIntegrationTest
+public class VertxHttpClientMetricsIT extends VertxHttpClientMetricsTest {
+
+}
